Add footer to model selector
