+++ crates/assistant/src/assistant_panel.rs
@@ -1,7 +1,6 @@
 use crate::{
     assistant_settings::{AssistantDockPosition, AssistantSettings},
     humanize_token_count,
-    model_selector2::ModelSelector2,
     prompt_library::open_prompt_library,
     slash_command::{
         default_command::DefaultSlashCommand,
@@ -2892,7 +2891,7 @@ impl Render for ContextEditorToolbarItem {
             });
         let right_side = h_flex()
             .gap_2()
-            .child(ModelSelector2::new(
+            .child(ModelSelector::new(
                 self.fs.clone(),
                 ButtonLike::new("active-model")
                     .style(ButtonStyle::Subtle)
